# docs/swagger.yaml
definitions:
...
  req.CreateOrgMemberReq:
    properties:
      avatar:
        description: 昵称
        type: string
      deptAndPositions:
        description: 部门id
        items:
          $ref: '#/definitions/req.DeptAndPositionReq'
        type: array
      email:
        description: 邮箱
        type: string
      empNo:
        description: 工号
        type: string
      loginName:
        description: 账号
        type: string
      name:
        description: 昵称
        type: string
      password:
        description: 密码
        type: string
      phoneNumber:
        description: 手机号
        type: string
      phoneRegion:
        description: 手机区号
        type: string
      roleIds:
        description: 角色id
        items:
          type: integer
        type: array
      status:
        description: 状态（1启用2禁用3离职, 默认启用） 忽视0
        type: integer
      weiboIds:
        description: 微博
        items:
          type: string
        type: array
    type: object
...
  req.DeptAndPositionReq:
    properties:
      departmentId:
        description: DepartmentId 部门ID(必填)
        type: integer
      positionId:
        description: PositionId 职级ID(必填 注意是Org内的OrgPositionId)
        type: integer
    type: object
...
  req.ExportAddressListReq:
    properties:
      authFilter:
        description: 是否开启权限过滤
        type: boolean
      departmentId:
        description: 部门id
        type: integer
      exportField:
        description: 导出字段(name,mobile,empNo,email,department,role,isLeader,statusChangeTime,createTime)
        items:
          type: string
        type: array
      isAllocate:
        description: 是否已分配部门（1已分配2未分配，默认全部）
        type: integer
      roleId:
        description: 角色id
        type: integer
      searchCode:
        description: 搜索字段
        type: string
      status:
        description: 是否禁用（1启用2禁用，默认全部）
        type: integer
    type: object
...
  req.GetInviteInfoReq:
    properties:
      inviteCode:
        type: string
    type: object
...
  req.IdReq:
    properties:
      id:
        description: ID
        type: integer
    type: object
...
  req.InviteUserData:
    properties:
      email:
        description: 邮箱
        type: string
      name:
        description: 姓名（再次邀请时不用传了）
        type: string
    type: object
  req.InviteUserListReq:
    properties:
      page:
        description: 页码（选填，不填为全部）
        type: integer
      size:
        description: 数量（选填，不填为全部）
        type: integer
    type: object
  req.InviteUserReq:
    properties:
      data:
        items:
          $ref: '#/definitions/req.InviteUserData'
        type: array
    type: object
...
  req.RemoveInviteUserReq:
    properties:
      ids:
        description: 要移除的id
        items:
          type: integer
        type: array
      isAll:
        description: 是否删除全部(1是，其余为否)
        type: integer
    type: object
  req.RemoveOrgMemberReq:
    properties:
      memberIds:
        description: 要移除的组织成员列表
        items:
          type: integer
        type: array
    type: object
...
  req.SearchUserReq:
    properties:
      email:
        type: string
    type: object
...
  req.UpdateOrgMemberReq:
    properties:
      adminGroup:
        description: 管理组。一个用户可以在多个管理组中。
        items:
          type: integer
        type: array
      avatar:
        description: 头像。可选
        type: string
      departmentIds:
        description: 用户所在的部门 ids. polaris 没有职级,因此无法使用 DeptAndPositions 字段
        items:
          type: integer
        type: array
      deptAndPositions:
        description: 部门/职级列表
        items:
          $ref: '#/definitions/req.DeptAndPositionReq'
        type: array
      email:
        description: 邮箱。可选
        type: string
      empNo:
        description: 工号。可选
        type: string
      name:
        description: 姓名。可选
        type: string
      phoneNumber:
        description: 手机号。可选
        type: string
      phoneRegion:
        description: 手机区号。可选，不传表示不更新该字段。
        type: string
      roleIds:
        description: 角色id
        items:
          type: integer
        type: array
      status:
        description: 状态（1启用2禁用3离职 选填）
        type: integer
      userId:
        description: UserId
        type: integer
      weiboIds:
        description: 微博。
        items:
          type: string
        type: array
    type: object
  req.UpdateOrgMemberStatusReq:
    properties:
      memberIds:
        description: 要修改的组织成员列表
        items:
          type: integer
        type: array
      status:
        description: 状态,  1可用,2禁用,3离职
        type: integer
    type: object
...
  req.UpdateUserInfoReq:
    properties:
      avatar:
        description: 用户头像
        type: string
      birthday:
        description: 生日
        type: string
      name:
        description: 姓名
        type: string
      sex:
        description: 性别
        type: integer
      updateFields:
        description: 变动的字段列表
        items:
          type: string
        type: array
    type: object
  req.UserListReq:
    properties:
      authFilter:
        description: 是否开启权限过滤
        type: boolean
      checkStatus:
        description: 审核状态。1待审核；2审核通过；3审核不通过
        items:
          type: integer
        type: array
      departmentId:
        description: 部门id
        type: integer
      isAllocate:
        description: 是否已分配部门（1已分配2未分配，默认全部）
        type: integer
      order:
        description: 排序字段 如"create_time desc,id asc"
        type: string
      page:
        description: 页码（选填，不填为全部）
        type: integer
      positionId:
        description: 职级ID
        type: integer
      roleId:
        description: 角色id
        type: integer
      searchCode:
        description: 搜索字段
        type: string
      size:
        description: 数量（选填，不填为全部）
        type: integer
      status:
        description: 是否禁用（1启用2禁用，默认全部）
        type: integer
    type: object
...
  resp.ExportAddressListResp:
    properties:
      url:
        type: string
    type: object
  resp.GetInviteCodeResp:
    properties:
      expire:
        type: integer
      inviteCode:
        type: string
    type: object
...
  resp.InviteUserInfo:
    properties:
      email:
        description: 邮箱
        type: string
      id:
        description: 用户id
        type: integer
      inviteTime:
        description: 邀请时间
        type: string
      isInvitedRecent:
        description: 是否24h内已邀请
        type: boolean
      name:
        description: 名称
        type: string
    type: object
  resp.InviteUserListResp:
    properties:
      list:
        description: 列表
        items:
          $ref: '#/definitions/resp.InviteUserInfo'
        type: array
      total:
        description: 总数
        type: integer
    type: object
  resp.InviteUserResp:
    properties:
      invalidEmail:
        description: 不符合规范的邮箱
        items:
          type: string
        type: array
      invitedEmail:
        description: 已邀请的邮箱
        items:
          type: string
        type: array
      isUserEmail:
        description: 已经是用户的邮箱
        items:
          type: string
        type: array
      successEmail:
        description: 成功的邮箱
        items:
          type: string
        type: array
    type: object
...
  resp.OrgMemberInfoReq:
    properties:
      adminGroupList:
        description: 用户所在的管理组
        items:
          $ref: '#/definitions/bo.ManageGroupInfoBo'
        type: array
      avatar:
        description: 用户头像
        type: string
      createTime:
        description: 创建时间
        type: string
      creator:
        description: 组织内用户创建者
        type: integer
      creatorName:
        description: 组织内用户创建者名称
        type: integer
      departmentList:
        description: 部门/职级信息
        items:
          $ref: '#/definitions/resp.UserDeptPositionData'
        type: array
      email:
        description: 邮箱
        type: string
      empNo:
        description: 工号
        type: string
      isCreator:
        description: 是否是组织创建人
        type: boolean
      isOwner:
        description: 是否是组织拥有者
        type: boolean
      loginName:
        description: LoginName 用户名
        type: string
      name:
        description: 姓名
        type: string
      namePy:
        description: 姓名拼音
        type: string
      phoneNumber:
        description: 手机号码
        type: string
      phoneRegion:
        description: 手机区号
        type: string
      roleList:
        description: 角色信息
        items:
          $ref: '#/definitions/resp.UserRoleData'
        type: array
      status:
        description: 状态1启用2禁用
        type: integer
      statusChangeTime:
        description: 禁用时间
        type: string
      userId:
        description: id
        type: integer
      weiboIds:
        description: 微博ID
        items:
          type: string
        type: array
    type: object
  resp.PersonalInfo:
    properties:
      avatar:
        description: 头像
        type: string
      birthday:
        description: 生日
        type: string
      createTime:
        description: 创建时间
        type: string
      departmentList:
        description: 部门/职级信息
        items:
          $ref: '#/definitions/resp.UserDeptPositionData'
        type: array
      email:
        description: 邮箱
        type: string
      emplId:
        description: 工号
        type: string
      functions:
        description: 权限
        items:
          type: string
        type: array
      id:
        description: 主键
        type: integer
      isOrgOwner:
        description: IsOrgOwner 是否是组织拥有者
        type: boolean
      isOutCollaborator:
        description: IsOutCollaborator 是否是外部协作人
        type: boolean
      isSubAdmin:
        description: IsSubAdmin 是否是子管理员
        type: boolean
      isSysAdmin:
        description: IsSysAdmin 是否是系统管理员
        type: boolean
      language:
        description: 语言
        type: string
      lastEditPwdTime:
        description: LastEditPwdTime
        type: string
      lastLoginIp:
        description: 上次登录ip
        type: string
      lastLoginTime:
        description: 上次登录时间
        type: string
      level:
        description: 付费等级
        type: integer
      levelName:
        description: 付费等级名
        type: string
      loginFailCount:
        description: 登录失败次数
        type: integer
      loginName:
        description: 登录名
        type: string
      loginNameEditCount:
        description: 登录名编辑次数
        type: integer
      mobile:
        description: 电话
        type: string
      motto:
        description: 座右铭
        type: string
      name:
        description: 名称
        type: string
      orgCode:
        description: 组织code
        type: string
      orgId:
        description: 组织id
        type: integer
      orgName:
        description: 组织名称
        type: string
      passwordSet:
        description: 密码是否设置过(1已设置0未设置)
        type: integer
      remindBindPhone:
        description: 是否需要提醒（1需要2不需要）
        type: integer
      rimanente:
        description: 剩余使用时长
        type: integer
      roleList:
        description: 角色信息
        items:
          $ref: '#/definitions/resp.UserRoleData'
        type: array
      sex:
        description: 性别
        type: integer
      sourceChannel:
        description: 来源
        type: string
      thirdName:
        description: 第三方名称
        type: string
      updateTime:
        description: 更新时间
        type: string
    type: object
...
  resp.SearchUserResp:
    properties:
      status:
        description: 搜索结果（1可邀请2已邀请3已添加）
        type: integer
      userInfo:
        $ref: '#/definitions/resp.OrgMemberInfoReq'
    type: object
...
  resp.UserListResp:
    properties:
      list:
        description: 列表
        items:
          $ref: '#/definitions/resp.OrgMemberInfoReq'
        type: array
      total:
        description: 总数
        type: integer
    type: object
...
  resp.UserManageAuthResp:
    properties:
      id:
        type: integer
      isOrgOwner:
        description: 是否是组织拥有者
        type: boolean
      isSubAdmin:
        description: 是否是子管理员
        type: boolean
      isSysAdmin:
        description: 是否是系统管理员
        type: boolean
    type: object
...
  resp.UserOrganizationListResp:
    properties:
      list:
        description: 用户组织列表
        items:
          $ref: '#/definitions/resp.UserOrganization'
        type: array
    type: object
  resp.UserRoleData:
    properties:
      roleId:
        description: 角色id
        type: integer
      roleName:
        description: 角色名称
        type: string
    type: object
  resp.UserStatResp:
    properties:
      allCount:
        description: 所有成员数量
        type: integer
      forbiddenCount:
        description: 已禁用成员数量
        type: integer
      resignedCount:
        description: 已离职成员数量
        type: integer
      unallocatedCount:
        description: 未分配成员数量
        type: integer
      unreceivedCount:
        description: 未接受邀请成员数量
        type: integer
      waitAuditCount:
        description: 待审核成员数量
        type: integer
    type: object
...
info:
  contact: {}
  title: Ying Sou Sou Apis
paths:
...
  /usercenter/api/v1/user/create:
    post:
      consumes:
      - application/json
      description: 新建成员接口
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/req.CreateOrgMemberReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: integer
        "400":
          description: ""
      security:
      - token: []
      summary: 新建成员
      tags:
      - 成员
...
  /usercenter/api/v1/user/exportAddressList:
    post:
      consumes:
      - application/json
      description: 导出通讯录接口
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/req.ExportAddressListReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.ExportAddressListResp'
        "400":
          description: ""
      security:
      - token: []
      summary: 导出通讯录
      tags:
      - 成员
  /usercenter/api/v1/user/get-user-manage-auth:
    get:
      consumes:
      - application/json
      description: 成员管理组权限
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.UserManageAuthResp'
        "400":
          description: ""
      security:
      - token: []
      summary: 成员管理组权限
      tags:
      - 成员权限
  /usercenter/api/v1/user/getInviteCode:
    get:
      consumes:
      - application/json
      description: 获取邀请码接口
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.GetInviteCodeResp'
        "400":
          description: ""
      security:
      - token: []
      summary: 获取邀请码
      tags:
      - 成员邀请
  /usercenter/api/v1/user/getInviteInfo:
    post:
      consumes:
      - application/json
      description: 获取邀请信息接口
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/req.GetInviteInfoReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: ParseBody
        "400":
          description: ""
      security:
      - token: []
      summary: 获取邀请信息
      tags:
      - 成员邀请
...
  /usercenter/api/v1/user/inviteUser:
    post:
      consumes:
      - application/json
      description: 邀请成员(再次邀请)接口
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/req.InviteUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.InviteUserResp'
        "400":
          description: ""
      security:
      - token: []
      summary: 邀请成员/再次邀请
      tags:
      - 成员邀请
  /usercenter/api/v1/user/inviteUserList:
    post:
      consumes:
      - application/json
      description: 未接受邀请成员列表接口
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/req.InviteUserListReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.InviteUserListResp'
        "400":
          description: ""
      security:
      - token: []
      summary: 未接受邀请成员列表
      tags:
      - 成员邀请
...
  /usercenter/api/v1/user/member-info:
    post:
      consumes:
      - application/json
      description: 根据ID获取成员详情接口
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/req.IdReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.OrgMemberInfoReq'
        "400":
          description: ""
      security:
      - token: []
      summary: 根据ID获取成员详情
      tags:
      - 成员
  /usercenter/api/v1/user/personalInfo:
    post:
      consumes:
      - application/json
      description: 获取当前成员信息接口
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.PersonalInfo'
        "400":
          description: ""
      security:
      - token: []
      summary: 获取当前成员信息
      tags:
      - 成员
...
  /usercenter/api/v1/user/removeInviteUser:
    post:
      consumes:
      - application/json
      description: 删除成员邀请列表接口
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/req.RemoveInviteUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: ""
        "400":
          description: ""
      security:
      - token: []
      summary: 删除成员邀请列表
      tags:
      - 成员邀请
  /usercenter/api/v1/user/removeUser:
    post:
      consumes:
      - application/json
      description: 移除成员接口
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/req.RemoveOrgMemberReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: ""
      security:
      - token: []
      summary: 移除成员
      tags:
      - 成员
...
  /usercenter/api/v1/user/searchUser:
    post:
      consumes:
      - application/json
      description: 邀请时搜索成员接口
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/req.SearchUserReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.SearchUserResp'
        "400":
          description: ""
      security:
      - token: []
      summary: 邀请时搜索成员
      tags:
      - 成员邀请
...
  /usercenter/api/v1/user/update:
    post:
      consumes:
      - application/json
      description: 更新成员接口
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/req.UpdateOrgMemberReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: ""
      security:
      - token: []
      summary: 更新成员
      tags:
      - 成员
...
  /usercenter/api/v1/user/updatePersonalInfo:
    post:
      consumes:
      - application/json
      description: 更新当前成员信息接口
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/req.UpdateUserInfoReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: ""
      security:
      - token: []
      summary: 更新当前成员信息
      tags:
      - 成员
  /usercenter/api/v1/user/updateUserStatus:
    post:
      consumes:
      - application/json
      description: 启用/禁用成员接口
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/req.UpdateOrgMemberStatusReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: boolean
        "400":
          description: ""
      security:
      - token: []
      summary: 启用/禁用成员
      tags:
      - 成员
  /usercenter/api/v1/user/userList:
    post:
      consumes:
      - application/json
      description: 成员列表接口
      parameters:
      - description: 入参
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/req.UserListReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.UserListResp'
        "400":
          description: ""
      security:
      - token: []
      summary: 成员列表
      tags:
      - 成员
  /usercenter/api/v1/user/userOrgList:
    post:
      consumes:
      - application/json
      description: 用户组织列表接口
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.UserOrganizationListResp'
        "400":
          description: ""
      security:
      - token: []
      summary: 用户组织列表
      tags:
      - 用户
  /usercenter/api/v1/user/userStat:
    post:
      consumes:
      - application/json
      description: 成员tab统计接口
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/resp.UserStatResp'
        "400":
          description: ""
      security:
      - token: []
      summary: 成员tab统计
      tags:
      - 成员
...
swagger: "2.0"